Data Retention and Onboarding

+++ backend/backend/app/services/temporal_service.py
@@ -20,10 +20,13 @@ from temporalio.exceptions import WorkflowAlreadyStartedError
 from temporalio.service import RPCError
 
 from backend.app.exceptions import HTTPException
+from backend.app.models.dataclasses.DeleteResponseParams import DeleteResponseParams
 from backend.app.models.dataclasses.ImportFormParams import ImportFormParams
 from backend.app.models.dataclasses.user_tokens import UserTokens
+from backend.app.utils.date_utils import get_formatted_date_from_str
 from backend.config import settings
 from common.configs.crypto import Crypto
+from common.models.standard_form import StandardFormResponse
 from common.utils.asyncio_run import asyncio_run
 
 
@@ -113,6 +116,41 @@ class TemporalService:
         except Exception as e:
             loguru.logger.error(e)
 
+    async def add_scheduled_job_for_deleting_response(
+        self, response: StandardFormResponse
+    ):
+        if not settings.schedular_settings.ENABLED:
+            return
+        expiration_date = get_formatted_date_from_str(response.expiration)
+        try:
+            await self.check_temporal_client_and_try_to_connect_if_not_connected()
+            await self.client.create_schedule(
+                "delete_response_" + str(response.response_id),
+                schedule=Schedule(
+                    action=ScheduleActionStartWorkflow(
+                        "delete_response_workflow",
+                        id="delete_response_" + response.response_id,
+                        arg=DeleteResponseParams(
+                            response_id=response.response_id,
+                        ),
+                        task_queue=settings.temporal_settings.worker_queue,
+                        retry_policy=RetryPolicy(maximum_attempts=4),
+                    ),
+                    spec=ScheduleSpec(
+                        cron_expressions=[
+                            f"{expiration_date.second} {expiration_date.minute} {expiration_date.hour} {expiration_date.day} {expiration_date.month} {expiration_date.isoweekday()} {expiration_date.year}"
+                        ]
+                    ),
+                ),
+            )
+        except ScheduleAlreadyRunningError as e:
+            loguru.logger.info(e)
+        except HTTPException as e:
+            if e.status_code != HTTPStatus.SERVICE_UNAVAILABLE:
+                loguru.logger.error(e)
+        except Exception as e:
+            loguru.logger.error(e)
+
     async def delete_form_import_schedule(
         self, workspace_id: PydanticObjectId, form_id: str
     ):
@@ -121,6 +159,23 @@ class TemporalService:
         try:
             await self.check_temporal_client_and_try_to_connect_if_not_connected()
             schedule_id = "import_" + str(workspace_id) + "_" + form_id
+            schedule_handle = self.client.get_schedule_handle(schedule_id)
+            await schedule_handle.delete()
+
+        except HTTPException:
+            pass
+        except RPCError as e:
+            loguru.logger.info(
+                "No schedule found for id:" + str(schedule_id) + " to delete"
+            )
+            pass
+
+    async def delete_response_delete_schedule(self, response_id: str):
+        if not settings.schedular_settings.ENABLED:
+            return
+        try:
+            await self.check_temporal_client_and_try_to_connect_if_not_connected()
+            schedule_id = "delete_response_" + response_id
             schedule_handle = self.client.get_schedule_handle(schedule_id)
             await schedule_handle.delete()
